context-graph/files/api/requirements.txt
# Knowledge Graph API - Dependencies

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
psycopg[binary,pool]==3.2.4
asyncpg==0.30.0
redis==5.2.1

# Vector database
qdrant-client==1.12.1

# LLM & Embeddings
anthropic==0.42.0
openai==1.59.3
tiktoken==0.8.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1

# Auth
PyJWT==2.9.0

# Observability
structlog==24.4.0
